+++ specs/examples/FunctionWithDependencies.ts
