+++ specs/examples/HigherOrderFunctionWithoutDependencies.ts
