+++ specs/examples/LoggerInterface.ts
