+++ specs/examples/MyService.ts
