+++ specs/examples/MyServiceClass.ts
