+++ specs/examples/MyServiceClassInterface.ts
