+++ specs/examples/MyServiceClassWithoutDependencies.ts
