+++ specs/examples/MyServiceInterface.ts
