+++ specs/examples/MyUseCase.ts
