+++ specs/examples/MyUseCaseInterface.ts
