+++ specs/examples/SayHelloType.ts
