+++ specs/examples/ServiceWithoutDependencyInterface.ts
