+++ specs/examples/sayHelloWorld.ts
